Reintroduce the inheritance in specs - need to use a generic to scope it
