+++ Specifications/Infrastructure/Services/Github/Webhooks/Handling/for_WebhookScheduler/when_scheduling_operations/and_more_webhooks_are_scheduled_after_the_first_batch_as_complete.cs
@@ -16,19 +16,14 @@ namespace Infrastructure.Services.Github.Webhooks.Handling.for_WebhookScheduler.
 {
     [Subject(typeof(IWebhookScheduler),"QueueWebhookEventForHandling")]
     public class and_more_webhooks_are_scheduled_after_the_first_batch_as_complete 
+        : given.a_webhook_scheduler_for<and_more_webhooks_are_scheduled_after_the_first_batch_as_complete>
     {
         static List<int> values;
         static List<given.Payload> payloads;
         static List<int> results;
-        static IWebhookScheduler scheduler;
-        static Mock<IWebhookProcessor> processor;
 
         Establish context = () => 
         {
-            var dependencies = given.dependencies.get();
-            scheduler = dependencies.scheduler;
-            processor = dependencies.processor;
-
             values = Enumerable.Range(0,10).ToList();
             results = new List<int>();
             payloads = new List<given.Payload>();
@@ -47,13 +42,13 @@ namespace Infrastructure.Services.Github.Webhooks.Handling.for_WebhookScheduler.
         {
             payloads.ForEach( _ => 
             {
-                var webhook = given.dependencies.build_webhook(_);
+                var webhook = given.a.webhook_from(_);
                 scheduler.QueueWebhookEventForHandling(webhook);
             });
             await Task.Delay(1000);
             payloads.ForEach(_ => 
             {
-                var webhook = given.dependencies.build_webhook(_);
+                var webhook = given.a.webhook_from(_);
                 scheduler.QueueWebhookEventForHandling(webhook);
             });
         };
